alog.py, record constrain 80 characters/line
The programm has an option to write a permanent ASCII record.  For
its display in a word processor like LibreOffice Writer assuming
either ISO A4 or US letter page, it is safer if there are only up
to 80 characters per line.  The comparison of two of these logs in
two terminals in simultaneous display equally is easier with this
constraint.