alog.py, edit presentation long file names in report
The previous approach to state the file names of structures
accessed in the permanent record fits well for short pattern
like from CCDC's CSD file

ABCDEFij.cif

with six upper-case Latin ASCII-7bit characters and optional
two digits (01, 02, etc) for a data set about a redetermination
or description of a new polymorphic form.

Working with .xyz of other origin may yield file names longer
than twelve characters in total; this then broke the layout of
the permanent record if seen e.g., in LibreOffice Writer.  By
redesign of the definitions, now up to 58 characters in total
may be display (aligned to the right hand side of the page) as
in the example of

BZAMID07-noH-really-quite-substantially-bit-longer0123.xyz

This commit only is about printing the file name into the report,
aRMSD may read and process files with file names consisting of
more than 58 characters (including file extension) successfully.