aplot.py, revert definition of angstrom
The original definition of Angstrom with \AA now is considered as
the better one because if there is a glyph for this in the family
at all, it will be consistent with the surounding characters.  The
excursion to a sans-serif math font just for this glyph complicates
the script's work more than necessary.