reformed initial presentation
Either README.md is, or/and README.org can be used like an
appetizer about what the program can do when accessing the
project on GitHub.  This purpose differs from an outline
how to use the program (which shall be addressed in a primer).
Thus re-using the illustrations about individual key steps
and display of the statistics plot.
For convenience, source file is README.org from which the .md
is exported in Emacs, and from which (about an intermediate
step on .tex) the analogue .pdf file.